Project 1 - Prompt Manager
I have worked with GPT to create a prompt manager. it has been a painful process as someone that wanted to do this before any udemy course.
This program has a .json library that also GPT for me modified and added all the awesome prompts from https://github.com/f/awesome-chatgpt-prompts/blob/main/prompts.csv

This is a BEFORE image. After this I am going to enroll on some prompt courses and redo in a couple of weeks.